LayoutTests/http/tests/xmlhttprequest/onerror-event-expected.txt

CONSOLE MESSAGE: XMLHttpRequest cannot load http://localhost:8000/xmlhttprequest/resources/access-control-basic-denied.cgi. Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
This test that the error event is fired for XMLHttpRequests

PASS: error event fired.
